Update ChampionManager.cs
매개변수 수정

see also : main

+++ Assets/Scripts/Character/ChampionManager.cs
@@ -196,7 +196,7 @@ public class ChampionManager : Controller
     }
 
     [PunRPC]
-    public void DamageCalculate(int damage)
+    public void DamageCalculate(float damage)
     {
         stat.Status.hp -= damage * (100 / (100 + stat.Status.defense));
     }
